0528 프로필 이미지 등록 기능 위치 변경
