+++ mypage/urls.py
@@ -1,6 +1,8 @@
+from django.conf.urls.static import static
 from django.urls import path
 from django.views.generic import RedirectView
 
+from config import settings
 from mypage.views import *
 
 app_name = 'mypage'
@@ -10,4 +12,7 @@ urlpatterns = [
     path('', index, name='index'),
     path('home/', RedirectView.as_view(url='/', permanent=True)),
 
-]
+    # 이미지 등록
+    path('profile/', ProfileUpload, name="profileUpload"),
+
+] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
